Connecting the projects together

+++ book-catalog-service/src/main/java/com/autodidact/bookcatalogservice/models/CatalogItem.java
@@ -7,9 +7,9 @@ public class CatalogItem {
 
 	private String name;
 	private String desc;
-	private int rating;
+	private double rating;
 
-	public CatalogItem(String name, String desc, int rating) {
+	public CatalogItem(String name, String desc, double rating) {
 		super();
 		this.name = name;
 		this.desc = desc;
